+++ notification/src/main/java/com/hubEleven/notification/ai/application/dto/MessageGenerationResponse.java
@@ -1,0 +1,11 @@
+package com.hubEleven.notification.ai.application.dto;
+
+//Ai가 산출한 최종 발송 시한 결과
+public record MessageGenerationResponse(
+        String finalDispatchDeadline,
+        String messageBody
+) {
+    public static MessageGenerationResponse of(String finalDispatchDeadline, String messageBody) {
+        return new MessageGenerationResponse(finalDispatchDeadline, messageBody);
+    }
+}
