+++ notification/src/main/java/com/hubEleven/notification/ai/domain/model/RequestStatus.java
@@ -1,0 +1,10 @@
+package com.hubEleven.notification.ai.domain.model;
+
+import lombok.Getter;
+
+@Getter
+public enum RequestStatus {
+    REQUESTED,
+    SUCCESS,
+    FAIL
+}
