+++ notification/src/main/java/com/hubEleven/notification/ai/infrastructure/client/dto/GeminiRequest.java
@@ -1,0 +1,11 @@
+package com.hubEleven.notification.ai.infrastructure.client.dto;
+
+import java.util.List;
+
+public record GeminiRequest(List<Content> contents) {
+    public static GeminiRequest fromPrompt(String prompt) {
+        return new GeminiRequest(List.of(new Content(List.of(new Part(prompt)))));
+    }
+    public record Content(List<Part> parts) {}
+    public record Part(String text) {}
+}
