+++ notification/src/main/java/com/hubEleven/notification/ai/infrastructure/configuration/AiProperties.java
@@ -1,0 +1,11 @@
+package com.hubEleven.notification.ai.infrastructure.configuration;
+
+import org.springframework.boot.context.properties.ConfigurationProperties;
+
+@ConfigurationProperties(prefix = "gemini")
+public record AiProperties(
+        Api api,
+        String model
+) {
+    public record Api(String key) {}
+}
