--- notification/src/main/java/com/hubEleven/notification/application/dto/MessageGenerationResponse.java
@@ -1,13 +1,0 @@
-package com.hubEleven.notification.application.dto;
-
-import jakarta.validation.constraints.NotBlank;
-
-//Ai가 산출한 최종 발송 시한 결과
-public record MessageGenerationResponse(
-        String finalDispatchDeadline,
-        String messageBody
-) {
-    public static MessageGenerationResponse of(String finalDispatchDeadline, String messageBody) {
-        return new MessageGenerationResponse(finalDispatchDeadline, messageBody);
-    }
-}
