--- notification/src/main/java/com/hubEleven/notification/domain/model/RequestStatus.java
@@ -1,10 +1,0 @@
-package com.hubEleven.notification.domain.model;
-
-import lombok.Getter;
-
-@Getter
-public enum RequestStatus {
-    REQUESTED,
-    SUCCESS,
-    FAIL
-}
