--- notification/src/main/java/com/hubEleven/notification/domain/repository/AiRequestLogRepository.java
@@ -1,13 +1,0 @@
-package com.hubEleven.notification.domain.repository;
-
-import com.hubEleven.notification.domain.model.AiRequestLog;
-
-import java.util.Optional;
-import java.util.UUID;
-
-public interface AiRequestLogRepository {
-    AiRequestLog save(AiRequestLog aiRequestLog);
-    Optional<AiRequestLog> findById(UUID id);
-    boolean existsByOrderId(UUID orderId);
-    Optional<AiRequestLog> findByOrderId(UUID orderId);
-}
