--- notification/src/main/java/com/hubEleven/notification/infrastructure/configuration/AiProperties.java
@@ -1,11 +1,0 @@
-package com.hubEleven.notification.infrastructure.configuration;
-
-import org.springframework.boot.context.properties.ConfigurationProperties;
-
-@ConfigurationProperties(prefix = "gemini")
-public record AiProperties(
-        Api api,
-        String model
-) {
-    public record Api(String key) {}
-}
